requirements.txt
celery==5.3.4
fastapi[all]==0.104.1
psycopg2-binary==2.9.9
pydantic==2.4.2
pydantic_core==2.10.1
python-dateutil==2.8.2
python-dotenv==1.0.0
redis==4.6.0
requests==2.31.0
SQLAlchemy==2.0.23
uvicorn==0.24.0.post1
